requirements.txt
phidata
qdrant-client
ollama
pypdf 
openai
fastapi
uvicorn